move 'page/page.js' to 'src/desktop.js' + minor reformatting
